pyproject.toml
[tool.poetry]
name = "bank-transaction-analysis"
version = "0.1.0"
description = ""
authors = ["Pavel Polezhaev <pavelpoleczaev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
python-dotenv = "^1.0.1"
pytest = "^8.3.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"


[tool.poetry.group.lint.dependencies]
mypy = "^1.11.0"
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
line_length = 119
